manga_translator_mvp/requirements.txt
gradio>=5.22.0
opencv-python>=4.10.0
Pillow>=10.4.0
numpy>=1.26.4
deep-translator>=1.11.4
rapidocr-onnxruntime>=1.4.4
google-genai>=1.11.0
